pyproject.toml
[project]
name = "orbit"
version = "0.1.0"
description = "AI-powered PR crisis command center with AGNTCY multi-agent system"
authors = [
    {name = "AGNTCY", email = "hello@agntcy.org"}
]
dependencies = [
    "agntcy-app-sdk==0.1.0",
    "a2a-sdk==0.2.5",
    "mcp>=1.10.0",
    "langgraph>=0.4.1",
    "langchain-openai>=0.3.16",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "openai>=1.3.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "dotenv>=0.9.9",
    "aiofiles>=23.2.1",
    "typing-extensions>=4.12.2",
    "watchfiles>=0.21.0",
]
requires-python = ">=3.12"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agents*", "common*", "gateway*"]
exclude = ["data*", "config*"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "RUF"]
ignore = ["E501"]  # Line too long

[tool.ruff.format]
quote-style = "double"
indent-style = "space"